+++ src/App.css
@@ -3,26 +3,21 @@
   height: 100vh;
   width: 100vw;
 }
-body[data-theme="light"] {
-  --bg-color: #fff;
-  --text-color: #000;
+
+.app.light {
+  background-color: #fff;
+  color: #333;
 }
 
-body[data-theme="dark"] {
-  --bg-color: #333;
-  --text-color: #fff;
-}
-
-.main-component {
-  background-color: var(--bg-color);
-  color: var(--text-color);
+.app.dark {
+  background-color: #333;
+  color: #fff;
 }
 
 .app-left {
   flex: 1;
   display: flex;
   flex-direction: column;
-  background-color: #f0f0f0;
 }
 
 .app-top {
@@ -37,7 +32,6 @@ body[data-theme="dark"] {
 
 .app-right {
   flex: 0 0 auto;
-  background-color: #ffffff;
   border-left: 1px solid #ddd;
   padding: 0;
   box-sizing: border-box;
@@ -50,6 +44,15 @@ body[data-theme="dark"] {
 
 .app-bottom.empty {
   background-color: transparent;
+}
+
+.app.light .button {
+  background-color: #000000ef;
+  color: #fff;
+}
+.app.app.dark .button {
+  background-color: #fff;
+  color: #333;
 }
 
 .loading {
