+++ src/components/DetailedCard/DetailedCard.css
@@ -1,5 +1,5 @@
 .detailed-card {
-  background-color: #ffffff;
+  /* background-color: #ffffff; */
   padding: 20px;
   border-radius: 10px;
   box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
@@ -31,7 +31,7 @@
 .detailed-card p {
   margin: 0;
   font-size: 18px;
-  color: #666666;
+  /* color: #666666; */
   text-align: center;
 }
