+++ src/components/ResultsComponent/ResultsComponent.css
@@ -12,9 +12,19 @@
   font-weight: bold;
   margin-top: 20px;
   text-align: center;
-  font-family: 'Arial', sans-serif;
+  font-family: "Arial", sans-serif;
   padding: 10px;
   background-color: #f4f4f4;
   border-radius: 5px;
   border: 1px solid #ddd;
+}
+
+.results.light {
+  background-color: white;
+  color: black;
+}
+
+.results.dark {
+  background-color: black;
+  color: white;
 }
