+++ src/components/ResultsComponent/ResultsComponent.tsx
@@ -1,5 +1,3 @@
-// ResultsComponent.tsx
-
 import React from "react";
 import "./ResultsComponent.css";
 import PokemonCard from "../PokemonCard/PokemonCard";
@@ -13,20 +11,22 @@ interface ResultsComponentProps {
 const ResultsComponent: React.FC<ResultsComponentProps> = ({
   pokemons,
   error,
-}) => (
-  <div className="results">
-    {error ? (
-      <p className="error-message">{error}</p> // Отображаем сообщение об ошибке
-    ) : pokemons.length > 0 ? (
-      pokemons.map((pokemon, index) => (
-        <div key={index}>
-          <PokemonCard pokemon={pokemon} />
-        </div>
-      ))
-    ) : (
-      <p className="no-results">No Pokemon found</p> // Сообщение, если покемоны не найдены
-    )}
-  </div>
-);
+}) => {
+  return (
+    <div className="results">
+      {error ? (
+        <p className="error-message">{error}</p> // Отображаем сообщение об ошибке
+      ) : pokemons.length > 0 ? (
+        pokemons.map((pokemon, index) => (
+          <div key={index}>
+            <PokemonCard pokemon={pokemon} />
+          </div>
+        ))
+      ) : (
+        <p className="no-results">No Pokemon found</p> // Сообщение, если покемоны не найдены
+      )}
+    </div>
+  );
+};
 
 export default ResultsComponent;
