+++ src/components/SearchComponent/SearchComponent.css
@@ -15,8 +15,12 @@
   max-width: 400px;
 }
 
+.search-form.search-form.dark ::placeholder {
+  color: #fff;
+}
+
 .search-button {
-  background-color: #2196f3;
+  background-color: #216bf3;
   color: #fff;
   border: none;
   padding: 12px 25px;
@@ -28,8 +32,11 @@
     transform 0.3s ease;
 }
 
+.search-form.light button {
+  background-color: #21adee;
+}
 .search-button:hover {
-  background-color: #1976d2;
+  background-color: #1976d2bd;
   transform: scale(1.05);
 }
 
@@ -38,7 +45,12 @@
   transform: scale(0.98);
 }
 
-.throw-error-button {
+.search-form.dark input {
+  background-color: #726f6f;
+  color: #000;
+}
+
+/* .throw-error-button {
   background-color: #f44336;
   color: #fff;
   border: none;
@@ -50,9 +62,9 @@
     background-color 0.3s ease,
     transform 0.3s ease;
   margin-top: 10px;
-}
+} */
 
-.throw-error-button:hover {
+/* .throw-error-button:hover {
   background-color: #d32f2f;
   transform: scale(1.05);
 }
@@ -60,4 +72,4 @@
 .throw-error-button:active {
   background-color: #b71c1c;
   transform: scale(0.98);
-}
+} */
